pyproject.toml
[project]
name = "seo-agent-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "ibis-framework[duckdb]>=10.4.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.1",
    "python-dotenv>=1.1.0",
    "tenacity>=9.0.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "pytest-mock>=3.14.0",
    "pytest>=8.3.5",
    "ruff>=0.11.2",
]
